Added antispam

+++ src/main/java/edu/sjsu/cs/listener/DiscordEventListener.java
@@ -10,12 +10,17 @@ import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
 import net.dv8tion.jda.api.sharding.ShardManager;
 import org.jetbrains.annotations.NotNull;
+import net.dv8tion.jda.api.entities.Message;
+import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import edu.sjsu.cs.ModeratorBot;
 
 import java.time.Duration;
+import java.util.HashMap;
 
 public class DiscordEventListener extends ListenerAdapter {
     public ModeratorBot bot;
+    private HashMap<String, Long> cooldowns;
+    private double cooldownPeriod;
     private String[] blacklistedWords = {
             "ugly",
             "stupid",
@@ -25,6 +30,8 @@ public class DiscordEventListener extends ListenerAdapter {
     // DiscordEventListener allows the bot to listen to events and messages sent to the bot
     public DiscordEventListener(ModeratorBot bot) {
         this.bot = bot;
+        cooldowns = new HashMap<>();
+        cooldownPeriod = 0;
     }
 
     // Loads all guilds to prevent errors
@@ -39,11 +46,12 @@ public class DiscordEventListener extends ListenerAdapter {
         Guild guild = jda.getGuildById("1234228780260917300");
         if (guild != null) {
             guild.updateCommands().addCommands(
-                    Commands.slash("hello", "Have the bot say hello to you"),
                     Commands.slash("warn", "Issues a warning to a member of the server")
                             .addOption(OptionType.USER, "user", "Warn this user", true)
                             .addOption(OptionType.STRING, "reason", "Reason for warning issue", true)
-                            .addOption(OptionType.INTEGER, "duration", "Optionally set timeout duration in seconds", false)
+                            .addOption(OptionType.INTEGER, "duration", "Optionally set timeout duration (seconds)", false),
+                    Commands.slash("cooldown", "Set the cooldown period (seconds)")
+                            .addOption(OptionType.NUMBER, "duration", "Set the cooldown duration (Set to 0 to disable)", true)
             ).queue();
         }
     }
@@ -51,11 +59,7 @@ public class DiscordEventListener extends ListenerAdapter {
     // Initiates events according to slash commands the bot receives
     @Override
     public void onSlashCommandInteraction(SlashCommandInteractionEvent event) {
-        if (event.getName().equals("hello")) {
-            event.reply("Hello " + event.getUser().getAsMention() + "!!!!")
-                    .setEphemeral(true);
-        }
-        else if (event.getName().equals("warn")) {
+        if (event.getName().equals("warn")) {
             if (event.getOption("duration") != null) {
                 String messageWithDuration = "**" + event.getUser().getEffectiveName() + "** has timed out **" +
                         event.getOption("user").getAsMember().getEffectiveName() + "** for **" +
@@ -80,14 +84,42 @@ public class DiscordEventListener extends ListenerAdapter {
                         .queue();
             }
         }
+        else if(event.getName().equals("cooldown")) {
+            cooldownPeriod = event.getOption("duration").getAsDouble();
+            event.reply("Cooldown has been successfully set")
+                    .setEphemeral(true)
+                    .queue();
+            event.getChannel().sendMessage("Cooldown period has been set to: **" +
+                    event.getOption("duration").getAsString() + "** seconds").queue();
+        }
     }
 
     @Override
     public void onMessageReceived(MessageReceivedEvent event) {
-        for (String word: blacklistedWords) {
-            if (event.getMessage().getContentRaw().equalsIgnoreCase(word)) {
-                event.getChannel().sendMessage("Please refrain from using offensive language").queue();
-                event.getMessage().delete().queue();
+        if (!event.getAuthor().isBot()) {
+            String userId = event.getAuthor().getId();
+
+            if (cooldowns.containsKey(userId)) {
+                long lastMessageTime = cooldowns.get(userId);
+                long currentTime = System.currentTimeMillis();
+                long timeSinceLastMessage = currentTime - lastMessageTime;
+
+                if (timeSinceLastMessage < cooldownPeriod * 1000) {
+                    event.getChannel().sendMessage("Please refrain from spamming, " +
+                            event.getAuthor().getEffectiveName()).queue();
+                    event.getMessage().delete().queue();
+                    return;
+                }
+            }
+
+            cooldowns.put(userId, System.currentTimeMillis());
+
+            for (String word : blacklistedWords) {
+                if (event.getMessage().getContentRaw().equalsIgnoreCase(word)) {
+                    event.getChannel().sendMessage("Please refrain from using offensive language, " +
+                            event.getAuthor().getEffectiveName()).queue();
+                    event.getMessage().delete().queue();
+                }
             }
         }
     }
